refactor: remove Cognee
The knowledge graph never paid for its latency. Every answer still had
to be grounded in the project's own files, the index had to be marked
stale on every single file write, and the round trip sat directly in
front of the user's reply.

What was the local fallback, buildLocalProjectContext, is now the whole
implementation in lib/project-context.ts. It was already self-contained.

Drops 642 lines, three routes the web client never called
(/memory/status, /memory/reindex, /memory/context), the indexing stage
of the GitHub import, and markProjectMemoryPendingSafely, which fired a
detached database write on every file create, update and delete.

If a knowledge graph is ever wanted again it belongs behind
getProjectContext's signature, not threaded through the write path.

github.ts also picks up the session memo and the content-table helper
here, since removing the indexing stage touches the same functions.
