+++ apps/web/src/lib/projects-client.ts
@@ -1,9 +1,8 @@
 export * from "./projects-client/files";
 export * from "./projects-client/generation";
-export * from "./projects-client/memory";
+export * from "./projects-client/chat";
 export * from "./projects-client/projects";
 export * from "./projects-client/quota";
 export * from "./projects-client/types";
-export * from "./projects-client/waitlist";
 export { AiProviderCreditError } from "./projects-client/http";
 export { UpstreamRateLimitError } from "./projects-client/http";
